market: rename `models/bid.ts`  to `models/market.ts`
