+++ as/cNFT/assembly/models/market.ts
